backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
email-validator
python-multipart
